pyproject.toml
[project]
name = "sunkream-api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2,<6.0",
    "django-environ>=0.13.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "sunkream_api.settings"
python_files = ["tests/test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
